Refactor PgPoolCreator producer and processor
The PgPoolCreator change will not break existing implementations.
Indeed, the contract now uses the parent Pool type, so returning PgPool is fine.

In the recorder, the pool configurator now returns the parent Pool type.
Then, in the processor, new bean configurators are set up to support injection of the different pools:

- bare pool
- vendor pool
- mutiny pool
- mutiny vendor pool